Completed the completion rate topology

+++ processing-service/src/main/java/com/mo/processing_service/entity/CompletionMetric.java
@@ -10,7 +10,7 @@ public record CompletionMetric(
         OffsetDateTime windowEnd,
         int playCount,
         int completeCount,
-        BigDecimal completionRate,
+        double completionRate,
         OffsetDateTime computedAt
 ) {
 }
